remove contactHelpers file, move createDmChannelsForNewContacts to modelBuilders
