+++ packages/shared/src/db/modelBuilders.ts
@@ -383,3 +383,34 @@ export function postFromDmPostActivityEvent(
 function getReceivedAtFromId(postId: string) {
   return api.udToDate(postId.split('/').pop() ?? postId);
 }
+
+export function createDmChannelsForNewContacts(
+  newContactMatches: [string, string][]
+): db.Channel[] {
+  const newContacts = newContactMatches.map(([phoneNumber, contactId]) => ({
+    id: contactId,
+    phoneNumber,
+  }));
+
+  return newContacts.map((contact) => ({
+    id: contact.id,
+    contactId: contact.id,
+    type: 'dm' as const,
+    currentUserIsMember: null,
+    currentUserIsHost: null,
+    isDmInvite: false,
+    isPending: false,
+    isNewMatchedContact: true,
+    title: '',
+    isNew: true,
+    members: [
+      {
+        chatId: contact.id,
+        contactId: contact.id,
+        contact: contact,
+        membershipType: 'channel' as const,
+      },
+    ],
+    timestamp: Date.now(),
+  }));
+}
